refactor: :truck: uiディレクトリをsharedディレクトリ内に移動
