+++ packages/app/src/components/buttons/GameResetButton.tsx
@@ -1,18 +1,18 @@
 import resetButtonImageUrl from "@/assets/reset.png";
-import { ImageButton } from "@/components/ui/ImageButton";
+import { useResetGame } from "@/game/hooks/useResetGame";
+import { useScene } from "@/game/hooks/useScene";
+import { ImageButton } from "@/shared/ui/ImageButton";
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from "@/components/ui/Tooltip";
-import { useResetGame } from "@/game/hooks/useResetGame";
-import { useScene } from "@/game/hooks/useScene";
+} from "@/shared/ui/Tooltip";
 import { preloadImages } from "@/shared/utils/helpers";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
-import { HelpOverlay } from "../ui/HelpOverlay";
-import { NeonText } from "../ui/NeonText";
+import { HelpOverlay } from "../../shared/ui/HelpOverlay";
+import { NeonText } from "../../shared/ui/NeonText";
 
 preloadImages(resetButtonImageUrl);
