+++ packages/app/src/components/panels/LogPanel.tsx
@@ -1,10 +1,10 @@
-import { Panel } from "@/components/ui/Panel";
 import { useGame } from "@/game/hooks/useGame";
+import { Panel } from "@/shared/ui/Panel";
 import { cn } from "@/shared/utils/helpers";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
+import { NeonText } from "../../shared/ui/NeonText";
 import { ResultLogText } from "../texts/ResultLogText";
-import { NeonText } from "../ui/NeonText";
 
 export const LogPanel: FC = () => {
   const { results } = useGame();
