+++ packages/app/src/components/scenes/ModeSelectScene.tsx
@@ -1,10 +1,10 @@
 import { cn } from "@/shared/utils/helpers";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
+import { AppBackground } from "../../shared/ui/AppBackground";
 import { MultiPlayButton } from "../buttons/MultiPlayButton";
 import { SinglePlayButton } from "../buttons/SinglePlayButton";
 import { ToggleLangButton } from "../buttons/ToggleLangButton";
-import { AppBackground } from "../ui/AppBackground";
 
 export const ModeSelectScene: FC = () => {
   const { t } = useTranslation();
