+++ packages/app/src/components/texts/GameStatusText.tsx
@@ -1,5 +1,5 @@
-import { NeonText } from "@/components/ui/NeonText";
 import { useGame } from "@/game/hooks/useGame";
+import { NeonText } from "@/shared/ui/NeonText";
 import { cn } from "@/shared/utils/helpers";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
