+++ packages/app/src/components/texts/ResultLogText.tsx
@@ -2,7 +2,7 @@ import { useGame } from "@/game/hooks/useGame";
 import type { GameResult } from "@/game/types/result";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
-import { NeonText } from "../ui/NeonText";
+import { NeonText } from "../../shared/ui/NeonText";
 
 export interface ResultLogTextProps {
   result: GameResult;
