+++ packages/app/src/shared/ui/AppBackground.tsx
