+++ packages/app/src/shared/ui/HelpOverlay.tsx
