+++ packages/app/src/shared/ui/Panel.tsx
